+++ R/MainFunction.R
@@ -72,7 +72,7 @@
 #' @importFrom Rfast "rowsums" "Order" "colVars" "rowmeans" "standardise" "sort_unique" "cora" "cova"
 #' @importFrom e1071 "matchClasses"
 #' @importFrom mvnfast "dmvn"
-#' @importFrom slam "as.simple_sparse_array"
+#' @importFrom slam "as.simple_sparse_array" "as.simple_triplet_matrix"
 #' @importFrom corpcor "make.positive.definite"
 #' @importFrom mclust "Mclust" "mclustBIC"
 #' @importFrom utils "memory.limit"